requirements.txt
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
requests==2.31.0
nbformat==5.9.2
jupyter==1.0.0